sort dept ascending and print to csv
print directly instead of creating report
test with full dataset

i haven't added an if to print only if at least one order in department but in this is not needed. can be added later.

used a relative column path instead of absolute since products column 1 has commas and is not parsing correctly. instead of writing a csv reader, we could use ready made tools.